backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
motor==3.4.0
python-dotenv==1.0.1
transformers==4.40.1
torch==2.3.0
Pillow==10.3.0
pandas==2.2.2
google-generativeai==0.5.2
python-multipart==0.0.9
httpx==0.27.0
